+++ tests/opener_test.py
@@ -3,11 +3,19 @@ import builtins
 from io import BytesIO
 from pathlib import Path
 from json import load
+from warnings import warn
 
 from unittest import mock
 import pytest
 from PIL import Image, ImageCms, UnidentifiedImageError
-from pillow_heif import register_heif_opener, get_cfg_options, reset_cfg_options, HeifBrand
+from pillow_heif import (
+    register_heif_opener,
+    get_cfg_options,
+    reset_cfg_options,
+    HeifBrand,
+    libheif_info,
+    HeifCompressionFormat,
+)
 
 
 register_heif_opener()
@@ -21,6 +29,9 @@ heif_images = [e for e in all_images if e["valid"]]
 heic_images = [e for e in heif_images if e["name"].endswith(".heic")]
 hif_images = [e for e in heif_images if e["name"].endswith(".hif")]
 avif_images = [e for e in heif_images if e["name"].endswith(".avif")]
+if not libheif_info()["en_de_coders"][HeifCompressionFormat.AV1.name]:
+    warn("Skipping tests for `AV1` format, no codecs support.")
+    avif_images.clear()
 
 
 @pytest.mark.parametrize("img_info", heif_images)